+++ README.md
@@ -1,11 +1,46 @@
-# Aspirant ![CI](https://github.com/aspirant-project/aspirant/actions/workflows/ci.yml/badge.svg) [![NuGet](https://img.shields.io/nuget/v/Aspirant?logo=nuget)](https://www.nuget.org/packages/Aspirant/)
+# Aspirant ![CI](https://github.com/aspirant-project/aspirant/actions/workflows/ci.yml/badge.svg)
 
 Aspirant is a set of extensions and experiments for [.NET Aspire](https://github.com/dotnet/aspire) App Host projects.
 
-### Installing Aspirant
+## Installing Aspirant packages
 
-Install the `Aspirant` package. Using the `dotnet` command line in the project directory:
+Aspirant consists of the following packages:
+
+| Package | NuGet | Description |
+| ------- | ----- | ----------- |
+| [Aspirant.Hosting](#aspiranthosting) | [![NuGet](https://img.shields.io/nuget/v/Aspirant.Hosting.svg)](https://www.nuget.org/packages/Aspirant.Hosting) | Base set of extensions for Aspire App Host projects that only depends on Aspire.Hosting. |
+| [Aspirant.Hosting.Yarp](#aspiranthostingyarp) | [![NuGet](https://img.shields.io/nuget/v/Aspirant.Hosting.Yarp.svg)](https://www.nuget.org/packages/Aspirant.Hosting.Yarp) | A YARP ingress resource for Aspire App Host projects. |
+| [Aspirant.Hosting.Testing](#aspiranthostingtesting) | [![NuGet](https://img.shields.io/nuget/v/Aspirant.Hosting.Testing.svg)](https://www.nuget.org/packages/Aspirant.Hosting.Testing) | A set of testing extension for test projects that use [Aspire.Hosting.Testing](https://nuget.org/packages/Aspire.Hosting.Testing). |
+| [Aspirant.Hosting.Testing.xunit](#aspiranthostingtestingxunit) | [![NuGet](https://img.shields.io/nuget/v/Aspirant.Hosting.Testing.xunit.svg)](https://www.nuget.org/packages/Aspirant.Hosting.Testing.xunit) | A set of testing extension for xUnit test projects that use [Aspire.Hosting.Testing](https://nuget.org/packages/Aspire.Hosting.Testing). |
+
+Install the Aspirant package you want into the relevant Aspire App Host or testing project, e.g. using the `dotnet` command line in the project directory:
 
 ```shell
-dotnet add package Aspirant
+dotnet add package Aspirant.Hosting
 ```
+
+## Aspirant.Hosting
+
+TODO: Add documentation for Aspirant.Hosting
+
+## Aspirant.Hosting.Yarp
+
+TODO: Add documentation for Aspirant.Hosting.Yarp
+
+## Aspirant.Hosting.Testing
+
+TODO: Add documentation for Aspirant.Hosting.Testing
+
+## Aspirant.Hosting.Testing.xunit
+
+TODO: Add documentation for Aspirant.Hosting.Testing.xunit
+
+## License
+
+Aspirant is licensed under the [MIT License](./LICENSE).
+
+## Contributing
+
+We welcome contributions to Aspirant. Please [log an issue](https://github.com/aspirant-project/aspirant/issues/new) to discuss your contribution before submitting a pull request.
+
+TODO: Add a contributors guide.
