+++ src/Aspirant.Hosting.Testing.xunit/Aspirant.Hosting.Testing.xunit.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+      <PackageReference Include="Aspire.Hosting.Testing" />
+      <PackageReference Include="xunit" />
+      <PackageReference Include="Microsoft.SourceLink.GitHub" PrivateAssets="All" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Aspirant.Hosting.Testing\Aspirant.Hosting.Testing.csproj" />
+  </ItemGroup>
+
+</Project>
