+++ src/Aspirant.Hosting.Testing.xunit/XUnitExtensions.cs
@@ -1,25 +1,18 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using Aspire.Hosting.Testing;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Xunit.Abstractions;
 
-namespace IntegrationTests.Infrastructure;
+namespace Aspirant.Hosting.Testing;
 
-internal static partial class DistributedApplicationTestFactory
+/// <summary>
+/// Extension methods for configuring xUnit output.
+/// </summary>
+public static partial class DistributedApplicationTestFactory
 {
-    ///// <summary>
-    ///// Creates an <see cref="IDistributedApplicationTestingBuilder"/> for the specified app host assembly and outputs logs to the provided <see cref="ITestOutputHelper"/>.
-    ///// </summary>
-    //public static async Task<IDistributedApplicationTestingBuilder> CreateAsync(string appHostAssemblyPath, ITestOutputHelper testOutputHelper)
-    //{
-    //    var builder = await CreateAsync(appHostAssemblyPath, new XUnitTextWriter(testOutputHelper));
-    //    builder.Services.AddSingleton<ILoggerProvider, XUnitLoggerProvider>();
-    //    builder.Services.AddSingleton(testOutputHelper);
-    //    return builder;
-    //}
-
     /// <summary>
     /// Writes <see cref="ILogger"/> messages and resource logs to the provided <see cref="ITestOutputHelper"/>.
     /// </summary>
