+++ src/Aspirant.Hosting.Testing.xunit/XUnitLogger.cs
@@ -5,7 +5,7 @@ using System.Text;
 using Microsoft.Extensions.Logging;
 using Xunit.Abstractions;
 
-namespace IntegrationTests.Infrastructure;
+namespace Aspirant.Hosting.Testing;
 
 /// <summary>
 /// An <see cref="ILogger"/> that writes log messages to an <see cref="ITestOutputHelper"/>.
