+++ src/Aspirant.Hosting.Testing.xunit/XUnitLoggerProvider.cs
@@ -2,19 +2,20 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using Microsoft.Extensions.Logging;
+using Xunit.Abstractions;
 
-namespace IntegrationTests.Infrastructure;
+namespace Aspirant.Hosting.Testing;
 
 /// <summary>
-/// A logger provider that stores logs in an <see cref="LoggerLogStore"/>.
+/// An <see cref="ILoggerProvider"/> that creates <see cref="ILogger"/> instances that output to the supplied <see cref="ITestOutputHelper"/>.
 /// </summary>
-internal class StoredLogsLoggerProvider(LoggerLogStore logStore) : ILoggerProvider
+internal class XUnitLoggerProvider(ITestOutputHelper testOutputHelper) : ILoggerProvider
 {
     private readonly LoggerExternalScopeProvider _scopeProvider = new();
 
     public ILogger CreateLogger(string categoryName)
     {
-        return new StoredLogsLogger(logStore, _scopeProvider, categoryName);
+        return new XUnitLogger(testOutputHelper, _scopeProvider, categoryName);
     }
 
     public void Dispose()
