+++ src/Aspirant.Hosting.Testing.xunit/XUnitTextWriter.cs
@@ -4,7 +4,7 @@
 using System.Text;
 using Xunit.Abstractions;
 
-namespace IntegrationTests.Infrastructure;
+namespace Aspirant.Hosting.Testing;
 
 /// <summary>
 /// A <see cref="TextWriter"/> that writes to an <see cref="ITestOutputHelper"/>.
