+++ src/Aspirant.Hosting.Testing/ResourceExtensions.cs
@@ -1,7 +1,6 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
+﻿using Aspire.Hosting.ApplicationModel;
 
-namespace IntegrationTests.Infrastructure;
+namespace Aspirant.Hosting.Testing;
 
 internal static class ResourceExtensions
 {
