+++ src/Aspirant.Hosting.Testing/StoredLogsLogger.cs
@@ -1,9 +1,6 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
+﻿using Microsoft.Extensions.Logging;
 
-using Microsoft.Extensions.Logging;
-
-namespace IntegrationTests.Infrastructure;
+namespace Aspirant.Hosting.Testing;
 
 /// <summary>
 /// A logger that stores logs in a <see cref="LoggerLogStore"/>.
