+++ src/Aspirant.Hosting/Aspirant.Hosting.csproj
@@ -6,7 +6,6 @@
     <Nullable>enable</Nullable>
     <Description>A set of useful extensions and experiments for .NET Aspire App Host projects.</Description>
     <PackageTags>aspire hosting</PackageTags>
-    <PackageReadmeFile>README.md</PackageReadmeFile>
   </PropertyGroup>
 
   <ItemGroup>
@@ -14,10 +13,6 @@
       <PackageReference Include="Microsoft.Extensions.ServiceDiscovery.Yarp" />
       <PackageReference Include="Yarp.ReverseProxy" />
       <PackageReference Include="Microsoft.SourceLink.GitHub" PrivateAssets="All" />
-  </ItemGroup>
-
-  <ItemGroup>
-      <None Include="../../README.md" Pack="true" PackagePath="\" />
   </ItemGroup>
 
 </Project>
