+++ src/Aspirant.Hosting/WaitForDependencies/WaitForDependenciesExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Concurrent;
 using System.Runtime.ExceptionServices;
+using Aspire.Hosting;
 using Aspire.Hosting.ApplicationModel;
 using Aspire.Hosting.Lifecycle;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
@@ -7,7 +8,7 @@ using Microsoft.Extensions.Logging;
 using Polly;
 using Polly.Retry;
 
-namespace Aspire.Hosting;
+namespace Aspirant.Hosting;
 
 /// <summary>
 /// Extension methods to wait for dependencies.
