+++ src/Directory.Build.targets
@@ -1,0 +1,7 @@
+<Project>
+
+  <ItemGroup>
+    <None Include="../../README.md" Pack="true" PackagePath="\" />
+  </ItemGroup>
+
+</Project>
